requirements.txt
# Faster-Whisper GUI Requirements
# Compatible with Python 3.9 - 3.12 (Python 3.14 may have compatibility issues)

PyQt5>=5.15.0
faster-whisper>=1.0.0
ctranslate2==4.4.0
tokenizers>=0.13.0
huggingface-hub>=0.21.0
onnxruntime>=1.14.0
psutil>=5.9.0  # Memory management and monitoring

# GPU SUPPORT REQUIREMENTS (Manual Installation Needed):
# 1. NVIDIA CUDA Toolkit 11.8 (System Driver)
# 2. NVIDIA cuDNN for CUDA 11.x (Copy DLLs to CUDA bin folder)
# Without these, the app will crash in GPU mode.

# Noise Reduction (Optional)
deepfilternet==0.5.6
torch==2.0.1
torchaudio==2.0.2
soundfile==0.13.1 # Required for torchaudio backend on Windows
librosa>=0.10.0 # Audio resampling for VAD preprocessing

# Document Generation
python-docx>=0.8.11  # For creating Word documents

# Build Tools
pyinstaller
